+++ zcash_client_backend/CHANGELOG.md
@@ -49,6 +49,8 @@ workspace.
   - `TransparentAddressMetadata::standalone_script`
   - `TransparentAddressMetadata::redeem_script`
   - `WalletTransparentOutput::with_known_input_size`
+  - `WalletTransparentOutput::account_id`
+  - `WalletTransparentOutput::recipient_key_scope`
   - `impl From<sapling_crypto::Note> for Note`
   - `impl From<orchard::Note> for Note`
 
@@ -57,6 +59,10 @@ workspace.
   `zcash_encoding 0.4`, `zcash_protocol 0.8`, `zcash_address 0.11`,
   `zcash_transparent 0.7`, `zcash_primitives 0.27`.
 - `zcash_client_backend::data_api`:
+  - Changes to the `InputSource` trait:
+    - The result types of `InputSource::get_unspent_transparent_output` and
+      `InputSource::get_unspent_transparent_outputs` have each changed; these
+      have reverted to returning `WalletTransparentOutput`.
   - Changes to the `WalletRead` trait:
     - `WalletRead::get_transparent_balances` now returns `TransparentBalances`
       (using `TransparentKeyOrigin`) to distinguish standalone transparent
@@ -111,6 +117,8 @@ workspace.
   - The associated type `ChangeStrategy::MetaSource` is now bounded on the newly
     added `MetaSource` type instead of `zcash_client_backend::data_api::InputSource`.
 - `zcash_client_backend::proposal`:
+  - `Proposal::single_step` and `Step::from_parts` now take transparent inputs
+    as `Vec<WalletTransparentOutput<()>>` (explicitly with no account ID).
   - `ProposalError` has added variants `PaymentAmountMissing` and
     `IncompatibleTxVersion`.
 - Migrated to `lightwallet-protocol v0.4.1`. This results in the following API
@@ -159,6 +167,11 @@ workspace.
   - `GapMetadata` has been moved behind the `transparent-inputs` feature flag,
     as it is only useful in the context of the wallet receiving transparent
     funds.
+  - `WalletTransparentOutput` has been refactored to convey information
+    equivalent to `WalletOutput`:
+    - It now has an `AccountId` generic parameter,
+    - `WalletTransparentOutput::from_parts` now takes `TransferType`,
+      `AccountId`, and `Option<TransparentKeyScope>` parameters.
   - `WalletTx::block_index` now has type `zcash_protocol::consensus::TxIndex`
     and the `WalletTx::new` function has been modified accordingly.
   - Renamed `TransparentAddressSource::Standalone` to
@@ -177,8 +190,10 @@ workspace.
   - `get_block_range_nullifiers`
 
 ### Removed
-- `zcash_client_backend::data_api::testing::transparent::GapLimits` (use
-  `zcash_keys::keys::transparent::GapLimits` instead).
+- `zcash_client_backend::data_api`:
+  - `WalletUtxo` (use `WalletTransparentOutput` instead).
+  - `testing::transparent::GapLimits` (use
+    `zcash_keys::keys::transparent::GapLimits` instead).
 
 ## [0.21.2] - 2026-03-10
 - The following APIs no longer crash in certain regtest mode configurations with